Fix crate detail flow and wire the offline cache
CrateViewModel called a non-existent no-arg getCrateDetails() and the
crate id was typed Int everywhere in the data layer while the UI already
passed a String, so the app didn't compile. Unified the id to String,
fixed the repo's offline fallback (it was retrying remote instead of
local), replaced the crash-on-touch TODO() local data source and the
never-initialized MoshiConverter with real Room-backed caching for both
crate summaries and crate details, and made CrateScreen render the
fetched CrateDetail instead of hardcoded example data. Also fixed the
Settings "Theme" item relaunching its own activity in a loop by making
it toggle night mode directly, and bumped the Kotlin/Java target to 11
so unit tests compile against mockito-kotlin.

+++ app/src/test/kotlin/io/github/andraantariksa/crates/feature_crates/data/source/remote/model/CratesIoDataSourceRemoteImplTest.kt
@@ -37,7 +37,7 @@ class CratesIoDataSourceRemoteImplTest {
     @Test
     fun `getCratesDetail should forward to cratesIoAPIService getCratesDetail`() = runBlocking {
         val crateDetail = CrateDetail.EXAMPLE
-        val crateId = crateDetail.crate.id.toInt()
+        val crateId = crateDetail.crate.id
         whenever(cratesIoAPIService.getCrateDetail(crateId)).thenReturn(crateDetail)
 
         val result = cratesIoDatasourceRemoteImpl.getCrateDetail(crateId)
